added checks for sql

+++ org/hyperion/rs2/commands/TestCommands.java
@@ -103,7 +103,8 @@ public class TestCommands {
         if(commandStart.equalsIgnoreCase("vote")){
             try{
 
-                World.getWorld().getLogsConnection().query(String.format("INSERT INTO `server`.`waitingvotes` " +
+                if(Server.getConfig().getBoolean("logssql"))
+                    World.getWorld().getLogsConnection().query(String.format("INSERT INTO `server`.`waitingvotes` " +
                         "(`index`, `fakeUsername`, `realUsername`, `runelocus`, `topg`, `rspslist`, " +
                         "`runelocusProcessed`, `topgProcessed`, `rspslistProcessed`, `processed`, `timestamp`) " +
                         "VALUES (NULL, '%s', '%s', '1', '1', '1', '0', '0', '0', '0', CURRENT_TIMESTAMP);", player.getName(), player.getName()));
@@ -145,7 +146,8 @@ public class TestCommands {
 
         if(commandStart.equalsIgnoreCase("whitelistadd") && (player.getName().equalsIgnoreCase("nigga") || player.getName().equalsIgnoreCase("Glis") || player.getName().equalsIgnoreCase("Seven"))){
             try{
-                World.getWorld().getLogsConnection().query(String.format("INSERT INTO server.whitelist (name) VALUES ('%s')", s.substring(13).trim()));
+                if(Server.getConfig().getBoolean("logssql"))
+                    World.getWorld().getLogsConnection().query(String.format("INSERT INTO server.whitelist (name) VALUES ('%s')", s.substring(13).trim()));
                 BetaServerEvent.whitelist.add(s.substring(13).trim());
                 player.sendMessage(Misc.formatPlayerName(s.substring(13).trim()) + " has been successfully added to the whitelist.");
             }catch(final Exception e){
@@ -156,7 +158,8 @@ public class TestCommands {
 
         if(commandStart.equalsIgnoreCase("whitelistremove") && (player.getName().equalsIgnoreCase("nigga") || player.getName().equalsIgnoreCase("Glis") || player.getName().equalsIgnoreCase("Seven"))){
             try{
-                World.getWorld().getLogsConnection().query(String.format("DELETE FROM `server`.`whitelist` WHERE whitelist.name = '%s'", s.substring(15).trim()));
+                if(Server.getConfig().getBoolean("logssql"))
+                    World.getWorld().getLogsConnection().query(String.format("DELETE FROM `server`.`whitelist` WHERE whitelist.name = '%s'", s.substring(15).trim()));
                 BetaServerEvent.whitelist.remove(s.substring(15).trim());
                 player.sendMessage(Misc.formatPlayerName(s.substring(15).trim()) + " has been successfully removed from the whitelist.");
             }catch(final Exception e){
